+++ dds-message-service/dds-message-client/src/main/java/com/icbc/dds/message/common/HeartbeatTask.java
@@ -7,7 +7,7 @@ import com.icbc.dds.api.exception.DDSRestRPCException;
 
 public class HeartbeatTask implements Runnable {
 
-	protected static final int HEARTBEAT_INTERVAL_MS = HeartbeatProperty.getHeartbeatIntervalMs();
+	protected static final long HEARTBEAT_INTERVAL_MS = HeartbeatProperty.getHeartbeatIntervalMs();
 	
 	private RpcClient rpcClient;
 	private Thread monitoredThread;
